requirements.txt
gradio
python-dotenv
langchain
langchain_community
langchain_openai
faiss-cpu